+++ NEWS.md
@@ -7,6 +7,7 @@
 * Allows filtering a GTFS feed by weekday and accounts for `calendar_dates.txt` info. Closes #103, #124 and #141
 * New sample GTFS.zip data for Berlin
 * New `compress` parameter in `gtfs2gps` function so that users can save the function outputs in compressed `.RDS` format.
+* New function `remove_by_route_id()` to remove GTFS data by route ids. Closes #180
 
 **Minor changes**
 * Fixed a bug in filter_day_period. Closes #96